refactor(viva-ms): extract into helper
