+++ services/viva/microservice/test/helpers/officers.test.ts
@@ -1,4 +1,4 @@
-import { createAdministrators } from '../../src/lambdas/syncOfficers';
+import createAdministrators from '../../src/helpers/createAdministrators';
 
 import type { CaseAdministrator } from '../../src/types/caseItem';
